Added Fast Fourier Transform Solver
Added two solvers that utilise the fact that M is a circulant matrix. FFT uses the Fast Fourier Transsform to first caclulate the eigen values of M, and the then uses it again to solve as each x is generated. FFTInv uses the Fast Fourier Transform method to cacualte the inverse of M, and then uses the inverse to solve for each x generated. Also added comments for main.